requirements.txt
pyTelegramBotAPI
schedule
python-dotenv
